Clean up some ISS code without functional changes
